% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faspic.R
\name{faspic}
\alias{faspic}
\title{Fitting ASPIC calling from R the ASPIC7 program}
\usage{
faspic(filename,t)
}
\arguments{
\item{filename}{Name of the ASPIC input file (.a7inp) or its path.}

\item{t}{integer. The time after which the execution of R is suspended either ASPIC program has finished the adjustment or not (see description). If it is less than the time required by the program to adjust ASPIC,  executable is closed before to obtain any results.}
}
\value{
The output files of  ASPIC7 program containing  the adjustment results.
}
\description{
The function opens the input file (.a7inp) in ASPIC7 program, once the input is read  ASPIC7g.exe run
ASPIC model based on such input information. Note that the adjustment is completed only if the time
fixed for such adjustment is greater than the required time for ASPIC7g.exe.
}
\details{
ASPIC programs are available on http://www.mhprager.com/aspic.html. Only work in windows.
}
\examples{
# Uncomment the following lines.
#faspic("aspic.a7inp",t=120)

}
\references{
Prager, M. (2016). User’s Guide for ASPIC Suite, version 7: A Stock–Production Model Incorporating Covariates and auxiliary programs. Portland, OR: Prager Consulting. Retrieved from http://www.mhprager.com/aspic7-guide.pdf.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Anxo Paz Cuña}
\item{Santiago Cerviño López}
\item{Maria Grazia Pennino}
}
}
