% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspic.R
\name{raspic}
\alias{raspic}
\title{Reading an ASPIC program output file}
\usage{
raspic (filename)
}
\arguments{
\item{filename}{Name of the ASPIC output file (.fit or .bot) or its path.}
}
\value{
A list containing the following components.
\itemize{
\item{errorcode:}{ Logistic. It takes the value 0 if there was no error in the adjustment or 1 if an error occurs.}
\item{convmsg:}{ The attained type of convergence.}
\item{states:}{ Data frame containing the following columns:}\itemize{
\item{time: }{Years of the input time series.}
\item{Fest: }{Estimated fishing mortality for each year.}
\item{B0est: }{Estimated biomass at the beginning of each year.}
\item{Best: }{Estimated average biomass for each year.}
\item{Catch: }{Observed catches for each year.}
\item{Cest: }{Estimated catches for each year.}
\item{SPest: }{Estimated surplus production for each year.}
\item{FFmsy: }{Estimated relative fishing mortality for each year (F_t/F_{msy}).}
\item{BBmsy: }{Estimated relative biomass (by B_{msy}) for each year (B_t/B_{msy}).}
\item{CPUEobs: }{Observed catches per unit of effort for each year.}
\item{CPUEest: }{Estimated catches per unit of effort for each year.}
\item{residuals: }{Residuals of the adjustment.}
\item{weight: }{Statistical weight of the observations. }
\item{BKest: }{Estimated relative biomass (by K) for each year (B_t/K).}}
\item{timespent:}{ Numeric. Time requires by ASPIC7 to adjust the model.}
\item{operation:}{ Data frame containing details of the type of adjustment carried out in ASPIC7.}\itemize{
\item{ASPIC.operation: }{Model used in the adjustment. Possible values are the logistic Schaefer model or the generalized Pella-Tomlinson model.}
\item{Objective.function: }{Estimation method. Possible values are "Least squares", "Least absolute values" or "Maximum likelihood".}
\item{Estimated.contrast.index: }{Estimated contrast index in the input series.}}
\item{estimates:}{ Estimates derived from the adjustment:}\itemize{
\item{MSY: }{Maximum sustainable yield (MSY).}
\item{Fmsy: }{Fishing mortality that gives MSY.}
\item{Bmsy: }{Biomass at MSY.}
\item{K: }{Maximum allowed population size.}
\item{phi: }{Relation between the MSY biomass and K (B_{msy}/K).}
\item{shape: }{Relation between the biomass at t=0 and the MSY biomass (B_0/B_{msy}).}
\item{Bly.Bmsy: }{Relation between the biomass in the last year plus one and the MSY biomass.}
\item{Fly.Fmsy: }{Relation between the fishing mortality in the last year and the MSY fishing mortality.}
\item{q: }{The catchability coefficient.}}}
}
\description{
This function reads ASPIC program output files (.fit or .bot), derived from the adjustment using ASPIC7g.exe program.
}
\examples{
# Uncomment the following lines.
#res<-raspic("aspic.fit")

# We can access to the previosly mentioned information:

# Adjustment info
#res$operation

# Estimations of yield and B/K relation:
#SPest<-res$states$SPest
#BKest<-res$states$BKest

# MSY, Fmsy and Bmsy
#res$estimates[1:3]

# Production curve
#plot(res$states$BKest,SPest)

}
\references{
Prager, M. (2016). User’s Guide for ASPIC Suite, version 7: A Stock–Production Model Incorporating Covariates and auxiliary programs. Portland, OR: Prager Consulting. Retrieved from http://www.mhprager.com/aspic7-guide.pdf.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Anxo Paz Cuña}
\item{Santiago Cerviño López}
\item{María Grazia Pennino}
}
}
